sam-service/requirements.txt
accelerate
fastapi
numpy
pillow
safetensors
sentencepiece
torch
torchvision
transformers
uvicorn[standard]
